Remove ".img" from kernel image file name.
When an ISO is created with the kernel image file name ending in ".img", the
ISO will not boot. In Qemu, the following error will be shown:

        FATAL: Int13 eltorito call with AX=4C00

This means a non-implemented BIOS entry was called. In VMware and on real
hardware there is a different error message:

        Cannot load disk image (invalid file)?

Removing the ".img" part of the file name solves this.