Added feature to exclude paths from the root filesystem.
To make the root filesystem as small as possible, it can be useful to exclude
certain files or directories. This is now possible by using the file
config/binary_rootfs/exclude, which contains a whitespace separated list of
file paths to exclude; wildcards are allowed. These paths are removed from the
chroot copy before the root filesystem is built.