Support list of items for conditionals in package lists
This patch allows giving a list of items for the conditionals in package
lists. For example, it's possible to specify some packages for more than one
architecture in the same conditional block:

 package01
 #if ARCHITECTURE i386 amd64
 package02
 package03
 #endif
 package04

where package01 and package04 are available for all architectures, but
package02 and package03 are available only for i386 and amd64 architectures.