Ensure packages in config/chroot_local-packages override all others.
This reverts config/chroot_local-packages to almost original semantics by
giving any package in chroot_local-packages a priority of 1001, allowing
it to "downgrade" packages and generally override the standard APT sources.

I think this follows the idea of this directory more and is likely to be
more useful with this setup - rebuilding a package simply to bump a version
number or keeping a custom apt_preferences(5) file around kinda sucks,
especially as it is often used as a temporary workaround.

Signed-off-by: Chris Lamb <chris@chris-lamb.co.uk>