Fix usb-hdd d-i installation.
The real fix was done in d-i; this patch change simply turns on
the right switches. In other words, you will need to use a very
new d-i for this to work, possibly with components from SVN.

Let's hold off on actually closing #510280 though, as it gives
the wrong impression.