Adding console-setup to standard package list.
console-setup is not essential per se, but without it, the preseed
boot parameters for keyboard layouts do not work. Those are needed
when we're going to use gfxboot. And that on the other hand, will
be pretty much standard, therefore including an additional package
in the standard list seems ok.