Allow to overload global functions with local functions stored in the config tree at scripts/functions.
live-helper 1.0.6 will be the last release of live-helper 1.x, which will
produce lenny 5.0.3 release images. However, when 5.0.4 is due, we'll need
to produce new images with slightly modified functions/release.sh files.
In order to acchieve that, we'll allow to have such a file overload the
functions shipped in live-helper.

This way, we can acchieve that:

  * live-helper itself doesn't need to be updated through stable updates
    anymore.
  * we still can do the required changes for a new point release.
  * we even ship the changes in the source tarballs automatically.