Moving 'default' hooks for minimal and stripped to example hooks, they will have to be enabled manually now (Closes: #627183).
Previously, there were a couple of hooks that were automatically run
for some image flavours. This came back from the old days when we
had to deconfigure gdm and enable kdm explicitly for kde images.

After lenny, this hack was not needed anymore, and there's not much
sense to keep the mechanism in place at all. Better, let's treat all
images the same way wrt/ hooks and don't do special automagics.

The only images that is of concern of are minimal and stripped
which we don't build by default anyway. This also stopps confusing
others.