Unifying config/chroot_archives and config/archives, now only using config/archives.
Previously, we had config/chroot_sources (or config/chroot_archives)
that allowed local configuration for third-party repositories but we
also had config/archives (or config/repositories) that did the very
same, but also alled to use 'variables' (e.g. @DISTRIBUTION@ or
@ARCHIVE_AREAS@ to be used) that got replaces automatically.

Now, we have only config/archives that can do both. This gets rid
of the unecessary double handling of the same thing in two places.