Make the config available to all hooks.
Previously, the config was bindmounted inside the chroot only for hooks
from the config dir, now it is done earlier so it is available for all
hooks.

Note that it is now also done if there are no hooks, but this is
probably enough of an exception case to not warrant the extra plumbing
required to detect the case (and doing an extra bindmount shouldn't
really hurt anyway...).