Dropping globale package lists for the desktop environments.
For the default images, we're maintaining their config trees in
git repositories now. So they're not needed anymore.

For people wanting to build e.g. a GNOME image in a 'quick&dirty'
way, in the past, they used something like 'lb config -p gnome'.

Since we do want to get away with all automagics like this where
we need to maintain the package list eventhough tasksel is already
doing that (and more), we're now needing people to create a package
list, e.g.

  # echo 'task-gnome-desktop' > config/package-lists/gnome-desktop.list.chroot

This is consistent with the package handling where we dropped the
'lb config --packages' in favour for the same mechanism for package lists.

With dropping global package lists finally, we'll have consistency
and same workflow for everything soon.