Removing minimal package list, not needed anymore.
live-boot pulls in eject and file through recommends,
live-config pulls in user-setup and sudo through recommends,
gnupg is pulled in as include during bootstrap on ubuntu.

We don't need these packages by default, and if users are
building with recommends disabled, then they have to make
sure these are pulled in too.

This is the more consistent handling than to explicitly
force all of these in all live images.