Removing bootstrap-flavour option (Closes: #685675).
live-build does too much magic that causes too much effort in
maintaining for stuff that hardly everyone uses.

Rather than 'translating' back and forth and maintaining compatibility
between options of different debootstrap and cdebootstrap versions (and
possibly any other in the future), it's is better to 'off-load' this to
the user.

For non-live specific options passed to the bootstrap tools, we'll
be introducing --debootstrap-options and --cdebootstrap-options
where users can specify their additional options passed to the
respective bootstrap tool, rather than us maintaining them in a
'hardcoded' and selected list.