Removing virtual-hdd image type.
Rather than a half broken image type, we'll bet our money on having proper
partman recipies support implemented for hdd image types.

Even if we don't make it for wheezy, it's better to not have virtual-hdd
in there and tell users to use the jessie version for proper hdd images.