Handling dist-upgrade right after bootstrap with copy of chroot_archives as bootstrap_archives in preparation for a future multi-archive bootstrap.
It is important for derivatives that the bootstrap cache is setup completely,
it is re-used in several places and thus cannot consist of the parent
distribution only.

Until there's a supportable way to bootstrap from multiple archives at the same time
(and multistrap doesn't work for that as discussed in the past), we're going to run
a lot of duplicated code here to massage the apt sources for the time being.