Dropping syslinux 6 symlinks from bootloader configs for wheezy.
In general there's no harm in having them as they are not used
on syslinux 4 distributions, however, but it's cosmetically nicer
and the path woudn't match anyway.

But.. since it turned out that we need the second copy of bootloaders
for wheezy anyway as wheezy and jessie are mutually exclusive
regarding paths where syslinux files are located, there's no
point in having them and confusing users with error messages
from cp -aL when stuff can't be dereferenced (which is ok).