Dropping unuseful information from bug-script (Closes: #774340).
Historically we tried to have all necessary information
in bug reports, such as version of certain packages used
during the build.

In reality, the only interesting package is debootstrap
on the host system - everything else is taken up from
the mirror from within the chroot and is noted in the
build log.

Users building non-chrooted will have to collect their
information by hand anyway, so it's better to remove
anything but debootstrap.